pyproject.toml
[tool.poetry]
name = "arcseamsbackend"
version = "0.1.0"
description = "Fashion design web app"
authors = ["Emmanuel <emmapraise@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.5"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
psycopg2-binary = "^2.9.5"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.2.2"
dj-database-url = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
